Rename package + use Groovy for Gradle config
