--- src/main/kotlin/Examples.kt
@@ -1,5 +1,0 @@
-import org.stellar.kotlinwalletsdk.Wallet
-
-fun main() {
-    // TODO: add examples
-}
