+++ src/main/kotlin/Main.kt
@@ -1,0 +1,3 @@
+fun main() {
+    // TODO: add examples
+}
