+++ src/main/kotlin/org/stellar/walletsdk/Wallet.kt
@@ -1,4 +1,4 @@
-package org.stellar.kotlinwalletsdk
+package org.stellar.walletsdk
 
 class Wallet(private val horizonUrl: String, private val horizonPassphrase: String) {
     // TODO: Wallet code goes here
