+++ src/test/kotlin/org/stellar/walletsdk/WalletTest.kt
@@ -1,4 +1,4 @@
-package org.stellar.kotlinwalletsdk
+package org.stellar.walletsdk
 
 class WalletTest {
     // TODO: Wallet tests go here
